DynamicRender2.0/pyproject.toml
[tool.poetry]
name = "DynamicRender2.0"
version = "0.1.0"
description = ""
authors = ["dunkirk-4 <lzxder@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.0"
pillow = "^9.1.0"
httpx = "^0.22.0"
colorlog = "^6.6.0"
qrcode = "^7.3.1"
emoji = "^1.7.0"
fonttools = "^4.31.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
matplotlib = "^3.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "douban"
url = "https://pypi.doubanio.com/simple/"
